my-app/requirements.txt
langgraph
langgraph-cli[inmem]
langchain
langchain_ollama
grpcio
grpcio-tools
protobuf
rich
dotenv
pydantic